@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,30 +27,34 @@ dependencies:
     sdk: flutter
 
   #IMPORT LIBRARIES
-  shared_preferences: ^2.0.5
-  toast: ^0.1.5
-  permission_handler: any
-  syncfusion_flutter_gauges: ^19.2.46
-  google_maps_flutter: ^0.5.30
-  location: ^3.0.2
-  geolocator: ^4.0.3
-  flutter_polyline_points: ^0.1.0
+  shared_preferences: ^2.0.6
+  #toast: ^0.1.5
+  fluttertoast: ^8.0.7
+  permission_handler: ^8.1.2
+  syncfusion_flutter_gauges: ^19.2.47
+  google_maps_flutter: ^2.0.6
+  location: ^4.3.0
+  geolocator: ^7.3.0
+  flutter_polyline_points: ^1.0.0
   pedometer: ^3.0.0
-  lottie: ^0.6.0
+  lottie: ^1.1.0
   flutter_rating_bar: ^4.0.0
   charts_flutter: ^0.11.0
-  volume: ^1.0.1
+  floor: ^1.1.0
+
   fl_chart: ^0.36.2
   intl: ^0.17.0
   stop_watch_timer: ^1.3.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  floor_generator: ^1.1.0
+  build_runner: ^2.0.6
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
