@@ -3,17 +3,17 @@ import 'dart:math' as math;
 import 'package:flutter/material.dart';
 
 class Wave extends StatefulWidget {
-  final double value;
+  final double? value;
   final Color color;
   final Axis direction;
   final int time;
 
   const Wave({
-    Key key,
-    @required this.value,
-    @required this.color,
-    @required this.direction,
-    @required this.time,
+    Key? key,
+    required this.value,
+    required this.color,
+    required this.direction,
+    required this.time,
   }) : super(key: key);
 
   @override
@@ -21,7 +21,7 @@ class Wave extends StatefulWidget {
 }
 
 class _WaveState extends State<Wave> with SingleTickerProviderStateMixin {
-  AnimationController _animationController;
+  late AnimationController _animationController;
 
   @override
   void initState() {
@@ -63,13 +63,13 @@ class _WaveState extends State<Wave> with SingleTickerProviderStateMixin {
 
 class _WaveClipper extends CustomClipper<Path> {
   final double animationValue;
-  final double value;
+  final double? value;
   final Axis direction;
 
   _WaveClipper({
-    @required this.animationValue,
-    @required this.value,
-    @required this.direction,
+    required this.animationValue,
+    required this.value,
+    required this.direction,
   });
 
   @override
@@ -97,7 +97,7 @@ class _WaveClipper extends CustomClipper<Path> {
       final waveHeight = (size.width / 20);
       final dx = math.sin((animationValue * 360 - i) % 360 * (math.pi / 180)) *
           waveHeight +
-          (size.width * value);
+          (size.width * value!);
       waveList.add(Offset(dx, i.toDouble()));
     }
     return waveList;
@@ -109,7 +109,7 @@ class _WaveClipper extends CustomClipper<Path> {
       final waveHeight = (size.height / 20);
       final dy = math.sin((animationValue * 360 - i) % 360 * (math.pi / 180)) *
           waveHeight +
-          (size.height - (size.height * value));
+          (size.height - (size.height * value!));
       waveList.add(Offset(i.toDouble(), dy));
     }
     return waveList;
