@@ -12,7 +12,7 @@ import 'package:run_tracker/utils/Debug.dart';
 import '../home/HomeScreen.dart';
 
 class WeeklyGoalSetScreen extends StatefulWidget {
-  const WeeklyGoalSetScreen({Key key}) : super(key: key);
+  const WeeklyGoalSetScreen({Key? key}) : super(key: key);
 
   @override
   _WeeklyGoalSetScreenState createState() => _WeeklyGoalSetScreenState();
@@ -22,14 +22,14 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
   bool KmSelected = true;
   bool MileSelected = false;
 
-  String unit;
+  String? unit;
   var weightKM = 1;
   var weightMILE = 1;
 
 
   @override
   Widget build(BuildContext context) {
-    unit = Languages.of(context).txtKM;
+    unit = Languages.of(context)!.txtKM;
     var fullHeight = MediaQuery.of(context).size.height;
     var fullWidth = MediaQuery.of(context).size.width;
     return Scaffold(
@@ -43,7 +43,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
             Container(
               margin: EdgeInsets.only(top: fullHeight * 0.15),
               child: Text(
-                Languages.of(context).txtYourWeeklyGoalIsReady,
+                Languages.of(context)!.txtYourWeeklyGoalIsReady,
                 style: TextStyle(
                     fontWeight: FontWeight.w700,
                     color: Colur.txt_white,
@@ -55,8 +55,8 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               child: Container(
                 margin: EdgeInsets.only(top: fullHeight * 0.07),
                 child: CustomTabBar(
-                    tab1: Languages.of(context).txtHeartHealth,
-                    tab2: Languages.of(context).txtDistance,
+                    tab1: Languages.of(context)!.txtHeartHealth,
+                    tab2: Languages.of(context)!.txtDistance,
                     forHeart: _forHeart(fullHeight, fullWidth),
                     forDistance: _forDistance(fullHeight)),
               ),
@@ -70,7 +70,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
     );
   }
 
-  Widget _forHeart([double fullHeight, double fullWidth]) {
+  Widget _forHeart(double fullHeight, double fullWidth) {
     return Expanded(
       child: Container(
         margin: EdgeInsets.only(top: fullHeight * 0.06),
@@ -100,7 +100,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
                   children: [
                     Container(
                       child: Text(
-                        Languages.of(context)
+                        Languages.of(context)!
                             .txt150MinBriskWalking
                             .toUpperCase(),
                         style: TextStyle(
@@ -113,7 +113,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
                       child: Padding(
                         padding: const EdgeInsets.only(bottom: 8.0, top: 5.0),
                         child: Text(
-                          Languages.of(context).txtPaceBetween9001500MinKm,
+                          Languages.of(context)!.txtPaceBetween9001500MinKm,
                           style: TextStyle(
                               fontWeight: FontWeight.w500,
                               color: Colur.txt_grey,
@@ -130,7 +130,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               alignment: Alignment.center,
               margin: EdgeInsets.symmetric(vertical: 15),
               child: Text(
-                Languages.of(context).txtOR,
+                Languages.of(context)!.txtOR,
                 style: TextStyle(
                     color: Colur.txt_grey,
                     fontSize: 20,
@@ -161,7 +161,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
                   children: [
                     Container(
                       child: Text(
-                        Languages.of(context).txt75MinRunning.toUpperCase(),
+                        Languages.of(context)!.txt75MinRunning.toUpperCase(),
                         style: TextStyle(
                             fontWeight: FontWeight.w700,
                             color: Colur.txt_white,
@@ -172,7 +172,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
                       child: Padding(
                         padding: const EdgeInsets.only(bottom: 8.0, top: 5.0),
                         child: Text(
-                          Languages.of(context).txtPaceover900MinKm,
+                          Languages.of(context)!.txtPaceover900MinKm,
                           style: TextStyle(
                               fontWeight: FontWeight.w500,
                               color: Colur.txt_grey,
@@ -201,7 +201,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
                     Padding(
                       padding: const EdgeInsets.only(left: 10.0),
                       child: Text(
-                        Languages.of(context)
+                        Languages.of(context)!
                             .txtYouCanCombineTheseTwoDescription,
                         maxLines: 2,
                         style: TextStyle(
@@ -220,7 +220,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
     );
   }
 
-  Widget _forDistance([double fullHeight]) {
+  Widget _forDistance(double fullHeight) {
     return Container(
       child: Column(
         children: [
@@ -234,7 +234,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
     );
   }
 
-  _distanceUnitTab([double fullHeight]) {
+  _distanceUnitTab(double fullHeight) {
     return Container(
       margin: EdgeInsets.only(top: fullHeight * 0.03),
       height: 60,
@@ -252,7 +252,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               setState(() {
                 KmSelected = true;
                 MileSelected = false;
-                unit = Languages.of(context).txtKM;
+                unit = Languages.of(context)!.txtKM;
               });
               Debug.printLog("$unit selected");
             },
@@ -260,7 +260,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               width: 100,
               child: Center(
                 child: Text(
-                  Languages.of(context).txtKM,
+                  Languages.of(context)!.txtKM,
                   style: TextStyle(
                       color: KmSelected ? Colors.white : Color(0xFF9195B6),
                       fontWeight: FontWeight.w500,
@@ -282,7 +282,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               setState(() {
                 KmSelected = false;
                 MileSelected = true;
-                unit = Languages.of(context).txtMILE;
+                unit = Languages.of(context)!.txtMILE;
                 Debug.printLog("$unit selected");
               });
             },
@@ -290,7 +290,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               width: 100,
               child: Center(
                 child: Text(
-                  Languages.of(context).txtMILE,
+                  Languages.of(context)!.txtMILE,
                   style: TextStyle(
                       color: MileSelected ? Colur.white : Color(0xFF9195B6),
                       fontWeight: FontWeight.w500,
@@ -313,7 +313,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
         height: 60,
         radius: 50.0,
         child: Text(
-          Languages.of(context).txtSetAsMyGoal.toUpperCase(),
+          Languages.of(context)!.txtSetAsMyGoal.toUpperCase(),
           maxLines: 1,
           overflow: TextOverflow.ellipsis,
           textAlign: TextAlign.center,
@@ -336,7 +336,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
     );
   }
 
-  _curpentinoPickerDesign([double fullHeight]) {
+  _curpentinoPickerDesign(double fullHeight) {
     return Container(
       child: Row(
         mainAxisAlignment: MainAxisAlignment.center,
@@ -359,7 +359,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               selectionOverlay: CupertinoPickerDefaultSelectionOverlay(background: Colur.transparent,),
               onSelectedItemChanged: (value) {
                 setState(() {
-                  if (unit == Languages.of(context).txtMILE) {
+                  if (unit == Languages.of(context)!.txtMILE) {
                     value += 1;
                     weightMILE = value;
                     Debug.printLog("$weightMILE $unit selected");
@@ -371,7 +371,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
                 });
               },
               itemExtent: 75.0,
-              children: unit == Languages.of(context).txtMILE
+              children: unit == Languages.of(context)!.txtMILE
                   ? List.generate(2155, (index) {
                       index += 1;
                       return Text(
@@ -398,7 +398,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               ? Container(
                   margin: EdgeInsets.only(left: 5),
                   child: Text(
-                    Languages.of(context).txtKM,
+                    Languages.of(context)!.txtKM,
                     style: TextStyle(
                         color: Colur.txt_white,
                         fontSize: 20,
@@ -408,7 +408,7 @@ class _WeeklyGoalSetScreenState extends State<WeeklyGoalSetScreen> {
               : Container(
                   margin: EdgeInsets.only(left: 5),
                   child: Text(
-                    Languages.of(context).txtMILE,
+                    Languages.of(context)!.txtMILE,
                     style: TextStyle(
                         color: Colur.txt_white,
                         fontSize: 20,
