@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 abstract class Languages {
-  static Languages of(BuildContext context) {
+  static Languages? of(BuildContext context) {
     return Localizations.of<Languages>(context, Languages);
   }
 
